Linear series - another linear roll table implementation
This is another variant of the linear roll table, but that
uses a constant, small amount memory for any size of table,
and all operations are O(1) when your table is one less
than power of two sized. Other sizes generate some rerolls.
It manages to do this by using a linear feedback shift
register.